refactor: rename file user-duplicated-email
