+++ src/core/user/interactor/create-user.interactor.ts
@@ -7,7 +7,7 @@ import { UserBirthdateInvalidError } from '../error/user-birthdate-invalid.error
 import { UserCellphoneInvalidError } from '../error/user-cellphone-invalid.error';
 import { UserCPFInvalidError } from '../error/user-cpf-invalid-error';
 import { UserDuplicatedCPFError } from '../error/user-duplicated-cpf.error';
-import { UserDuplicatedEmailError } from '../error/user-duplicated-emaild.error';
+import { UserDuplicatedEmailError } from '../error/user-duplicated-email.error';
 import { UserEmailInvalidError } from '../error/user-email-invalid.error';
 import { UserNameInvalidError } from '../error/user-name-invalid.error';
 import type { UserRepository } from '../user.repository';
